feat: :sparkles: Creat registerPet function
The function registerPet is created and registers a pet requesting data (name, species, age, weight and health). It validates that all fields are correct and, after a delay of 2 seconds, adds the pet to a pet array and displays a success message. (A bug is found when trying to display the message `“✅ Pet ${name} successfully registered.`”)